+++ .flake8
@@ -1,3 +1,4 @@
 [flake8]
+ignore = E302
 max-line-length = 121
 exclude = .git, __pycache__, venv, .venv
