+++ scr/api_client.py
@@ -1,5 +1,4 @@
 # api_client.py
-from dataclasses import dataclass
 import requests
 import time
 import logging
@@ -9,8 +8,11 @@ logging.basicConfig(level=logging.INFO)
 
 
 class Parser(ABC):
+    """Абстрактный класс для загрузки вакансий из API."""
+
     @abstractmethod
     def load_vacancies(self, keyword):
+        """Загружает вакансии по ключевому слову."""
         pass
 
 
@@ -20,12 +22,16 @@ class HH(Parser):
     """
 
     def __init__(self):
+        """Инициализирует параметры запроса к API."""
+
         self.url = 'https://api.hh.ru/vacancies'
         self.headers = {'User-Agent': 'HH-User-Agent'}
         self.params = {'text': '', 'page': 0, 'per_page': 100}
         self.vacancies = []
 
     def load_vacancies(self, keyword):
+        """Загружает вакансии с HeadHunter по ключевому слову."""
+
         self.params['text'] = keyword
         self.params['page'] = 0
         while self.params['page'] < 20:
@@ -45,78 +51,3 @@ class HH(Parser):
             except requests.RequestException as e:
                 logging.error(f"Ошибка запроса: {e}")
                 break
-
-
-@dataclass
-class Vacancy:
-    title: str
-    link: str
-    salary: str
-    description: str
-    company: str = "Не указана"
-    city: str = "Не указан"
-    experience: str = "Не указан"
-
-    def __post_init__(self):
-        if not self.title:
-            raise ValueError("Название вакансии не может быть пустым.")
-        if not self.link:
-            raise ValueError("Ссылка на вакансию не может быть пустой.")
-
-    @staticmethod
-    def parse_salary(salary):
-        """Обрабатывает зарплату из API HH"""
-        if isinstance(salary, dict):
-            salary_from = salary.get('from', 'Не указано')
-            salary_to = salary.get('to', 'Не указано')
-            currency = salary.get('currency', '')
-            return f"{salary_from} - {salary_to} {currency}".strip()
-        return salary if salary else "Зарплата не указана"
-
-    def __lt__(self, other):
-        """Сравнение вакансий по зарплате (меньше)"""
-        if not isinstance(self.salary, (int, float)) or not isinstance(other.salary, (int, float)):
-            return False
-        return self.salary < other.salary
-
-    def __gt__(self, other):
-        """Сравнение вакансий по зарплате (больше)"""
-        if not isinstance(self.salary, (int, float)) or not isinstance(other.salary, (int, float)):
-            return False
-        return self.salary > other.salary
-
-    def __eq__(self, other):
-        """Сравнение вакансий по зарплате (равно)"""
-        if not isinstance(self.salary, (int, float)) or not isinstance(other.salary, (int, float)):
-            return False
-        return self.salary == other.salary
-
-    def __str__(self):
-        return (f"Название: {self.title}\n"
-                f"Ссылка: {self.link}\n"
-                f"Зарплата: {self.salary}\n"
-                f"Описание: {self.description[:100] + '...' if self.description else 'Описание отсутствует'}\n")
-
-    # def __repr__(self):
-    #     """
-    #     Представление объекта вакансии в виде строки.
-    #     :return: Строковое представление вакансии.
-    #     """
-    #     return (f"Vacancy(title={self.title}, link={self.link}, "
-    #             f"salary={self.salary}, description={self.description})")
-    #
-# if __name__ == "__main__":
-#     # Создаем несколько вакансий
-#     vacancy1 = Vacancy("Python Developer", "http://example.com/python", 120000, "Опыт работы от 3 лет.")
-#     vacancy2 = Vacancy("Java Developer", "http://example.com/java", 100000, "Опыт работы от 2 лет.")
-#     vacancy3 = Vacancy("Data Scientist", "http://example.com/data", "Зарплата не указана", "Знание Python и SQL.")
-#
-#     # Сравниваем вакансии по зарплате
-#     print(f"Python Developer > Java Developer: {vacancy1 > vacancy2}")  # True
-#     print(f"Java Developer < Python Developer: {vacancy2 < vacancy1}")  # True
-#     print(f"Data Scientist == Python Developer: {vacancy3 == vacancy1}")  # False
-#
-#     # Выводим информацию о вакансиях
-#     print(vacancy1)
-#     print(vacancy2)
-#     print(vacancy3)
