+++ scr/main.py
@@ -1,6 +1,7 @@
 # main.py
-from api_client import HH, Vacancy
+from api_client import HH
 from storage import JSONStorage
+from vacancy import Vacancy  # Теперь импорт из vacancy.py
 
 
 def get_user_input(prompt, convert_func=str):
