+++ scr/storage.py
@@ -5,6 +5,7 @@ from abc import ABC, abstractmethod
 
 
 class Storage(ABC):
+    """Абстрактный класс для хранилища вакансий."""
     @abstractmethod
     def add_vacancy(self, vacancy):
         """Добавляет вакансию в хранилище"""
@@ -22,6 +23,8 @@ class Storage(ABC):
 
 
 class JSONStorage(Storage):
+    """Класс для работы с JSON-файлом как хранилищем вакансий."""
+
     def __init__(self, filename):
         """Инициализация хранилища JSON-файла"""
         current_dir = os.path.dirname(__file__)
@@ -38,13 +41,19 @@ class JSONStorage(Storage):
                 json.dump([], file)
 
     def add_vacancy(self, vacancy):
-        """Добавляет вакансию в JSON-файл"""
+        """Добавляет вакансию в JSON-файл, избегая дублирования"""
         if not vacancy.get("name") or not vacancy.get("alternate_url"):
             return
 
         try:
             with open(self.filename, 'r+', encoding='utf-8') as file:
                 vacancies = json.load(file)
+
+                # Проверяем, есть ли такая вакансия уже в файле
+                if any(v.get("name") == vacancy["name"] and v.get("company") == vacancy["company"] and v.get("city") ==
+                       vacancy["city"] for v in vacancies):
+                    return
+
                 vacancies.append(vacancy)
                 file.seek(0)
                 json.dump(vacancies, file, ensure_ascii=False, indent=4)
@@ -94,48 +103,3 @@ class JSONStorage(Storage):
 
     def __exit__(self, exc_type, exc_value, traceback):
         pass
-
-# if __name__ == "__main__":
-#     # Создаем хранилище
-#     storage = JSONStorage('vacancies.json')
-#
-#     # Пример данных из hh.ru
-#     hh_data = {
-#         "items": [
-#             {
-#                 "id": "93353083",
-#                 "name": "Тестировщик комфорта квартир",
-#                 "salary": {"from": 350000, "to": 450000, "currency": "RUB", "gross": False},
-#                 "alternate_url": "https://hh.ru/vacancy/93353083",
-#                 "snippet": {
-#                     "requirement": "Занимать активную жизненную позицию, уметь активно танцевать и громко петь...",
-#                     "responsibility": "Оценивать вид из окна: встречать рассветы на кухне..."
-#                 }
-#             },
-#             {
-#                 "id": "92223756",
-#                 "name": "Удаленный диспетчер чатов (в Яндекс)",
-#                 "salary": {"from": 30000, "to": 44000, "currency": "RUB", "gross": True},
-#                 "alternate_url": "https://hh.ru/vacancy/92223756",
-#                 "snippet": {
-#                     "requirement": "Обязательное знание русского языка...",
-#                     "responsibility": "Обработка входящих сообщений..."
-#                 }
-#             }
-#         ]
-#     }
-#
-#     # Добавляем вакансии в хранилище
-#     for item in hh_data['items']:
-#         storage.add_vacancy(item)
-#
-#     # Получаем вакансии с зарплатой от 30000
-#     print("Вакансии с зарплатой от 40000:")
-#     for vacancy in storage.get_vacancies({'salary_from': 30000}):
-#         print(vacancy)
-#
-#     # Удаляем вакансию по ID
-#     storage.delete_vacancy('93353083')
-#     print("Вакансии после удаления:")
-#     for vacancy in storage.get_vacancies({}):
-#         print(vacancy)
